Command refactor
